+++ lib/i18n/tasks/command/commands/eq_base.rb
@@ -5,9 +5,9 @@ module I18n::Tasks
         include Command::Collection
 
         cmd :eq_base,
-            args: '[locale ...]',
+            pos:  '[locale ...]',
             desc: t('i18n_tasks.cmd.desc.eq_base'),
-            opt:  cmd_opts(:locales, :out_format)
+            args: [:locales, :out_format]
 
         def eq_base(opt = {})
           print_forest i18n.eq_base_keys(opt), opt, :eq_base_keys