+++ lib/i18n/tasks/command/commands/meta.rb
@@ -5,7 +5,7 @@ module I18n::Tasks
         include Command::Collection
 
         cmd :config,
-            args: '[section ...]',
+            pos:  '[section ...]',
             desc: t('i18n_tasks.cmd.desc.config')
 
         def config(opts = {})