+++ lib/i18n/tasks/command/commands/missing.rb
@@ -4,26 +4,31 @@ module I18n::Tasks
       module Missing
         include Command::Collection
 
-        enum_list_opt :missing_types,
-                      '-t',
-                      '--types',
-                      I18n::Tasks::MissingKeys.missing_keys_types,
-                      proc { |valid| I18n.t('i18n_tasks.cmd.args.desc.missing_types', valid: valid) },
-                      proc { |invalid, valid| I18n.t('i18n_tasks.cmd.errors.invalid_missing_type',
-                                                invalid: invalid * ', ', valid: valid * ', ', count: invalid.length) }
+        missing_types = I18n::Tasks::MissingKeys.missing_keys_types
+        arg :missing_types,
+            '-t',
+            "--types #{missing_types * ','}",
+            Array,
+            t('i18n_tasks.cmd.args.desc.missing_types', valid: missing_types * ', '),
+            parser: OptionParsers::Enum::ListParser.new(
+                missing_types,
+                proc { |invalid, valid| I18n.t('i18n_tasks.cmd.errors.invalid_missing_type',
+                                               invalid: invalid * ', ', valid: valid * ', ', count: invalid.length) })
+
+
         cmd :missing,
-            args: '[locale ...]',
+            pos:  '[locale ...]',
             desc: t('i18n_tasks.cmd.desc.missing'),
-            opt:  cmd_opts(:locales, :out_format, :missing_types)
+            args: [:locales, :out_format, :missing_types]
 
         def missing(opt = {})
           print_forest i18n.missing_keys(opt), opt, :missing_keys
         end
 
         cmd :translate_missing,
-            args: '[locale ...]',
+            pos:  '[locale ...]',
             desc: t('i18n_tasks.cmd.desc.translate_missing'),
-            opt:  cmd_opts(:locales, :locale_to_translate_from) << cmd_opt(:out_format)[1..-1]
+            args: [:locales, :locale_to_translate_from, arg(:out_format).from(1)]
 
         def translate_missing(opt = {})
           missing    = i18n.missing_diff_forest opt[:locales], opt[:from]
@@ -34,9 +39,9 @@ module I18n::Tasks
         end
 
         cmd :add_missing,
-            args: '[locale ...]',
+            pos:  '[locale ...]',
             desc: t('i18n_tasks.cmd.desc.add_missing'),
-            opt:  cmd_opts(:locales, :out_format) << (cmd_opt(:value).dup << {default: '%{value_or_human_key}'})
+            args: [:locales, :out_format, arg(:value) + [{default: '%{value_or_human_key}'}]]
 
         def add_missing(opt = {})
           forest = i18n.missing_keys(opt).set_each_value!(opt[:value])